refactor: rename Offense to Build, Defense to Protect, and flatten route structure
- Rename sidebar navigation from Offense to Build and Defense to Protect
- Update route paths from /offense to /build for main navigation
- Flatten route structure: move all nested routes to root level
  - /defense/* → root level (insights, testing, janitors)
  - /context/* → root level (learn, calls)
- Update sidebar icons: Map → Blocks for Build, Shield → ShieldCheck for Protect
- Update all test references from nav-offense to nav-build
- Remove duplicate nested page directories
