+++ src/app/w/[slug]/learn/components/LearnChatArea.tsx
